Can challenge several time the same domain
We can have several TXT record with the same name so no need to delete
the already present record.

Indeed, when challenging stuff.com and *.stuff.com, the challenge is
done against _acme-challenge.stuff.com both time, so we must have both
validation

To achieve this, we try a first time to create the entry. If it already
exists, we retrieve the `rrset_values` before deleting the entry, and
recreate it with the old `rrset_values` with the new certbot validation